Refactor: 重构游戏计时系统和 timeline (#312)
* Refactor: 打包 timeline 上下文到结构体

* Refactor: 更新 TimerManager，使其每帧传出高精度的 deltaTime

* Refactor: 使 timeline 接受 dspTime，而非 audioSource.time

* Update: 添加两个 debug 防止以后卡顿找不到原因

* Fix: 计算逻辑错误

* Refactor: 用只读接口重构 TimelineContext*

* Timeline 运行时的更新逻辑仍待完善

* Update: 将 TimerManager 恢复原状，添加 SmoothDspTimer

* Update: 减少嵌套，更新调用逻辑

* Updte: 优化代码

* Fix

* Update: 由外部调用 timeline 更新，拆分提供 OnPlayingUpdate 和 JumpTo 方法区分

* Update: 修改 SmoothDspTimer 命名空间

* Update: 添加 DSPTimer Reset 方法以在暂停后忽略暂停间隔，更新到按每秒消除误差，移除 TODO

* Update: IClip.OnTimeSkip

* Update SmoothDspTimer.cs

* Refactor: 使 NoteClips 支持 Skip

* Update: 暂停/继续音游时正确刷新时间轴计时

* Fix: 计算错误

* Update: 将 SmoothDspTimer 和相关调用拆分

+++ Cyan-Stars/Assets/Scripts/Gameplay/MusicGame/Input/Receiver/KeyboardInputReceiver.cs
@@ -28,7 +28,7 @@ namespace CyanStars.Gameplay.MusicGame
             GameRoot.Timer.UpdateTimer.Remove(OnUpdate);
         }
 
-        private void OnUpdate(float deltaTime,object userdata)
+        private void OnUpdate(double deltaTime,object userdata)
         {
             foreach (InputMapData.Item item in InputMapData.Items)
             {
